factor out commands
