+++ packages/compass-e2e-tests/helpers/commands/index.ts
@@ -33,3 +33,8 @@ export * from './run-find';
 export * from './export-to-language';
 export * from './get-active-tab-namespace';
 export * from './click-parent';
+export * from './get-connect-form-state';
+export * from './set-connect-form-state';
+export * from './navigate-to-connect-tab';
+export * from './expand-connect-form-options';
+export * from './reset-connect-form';
